QCamera2: Fix slow-shutter feature with CameraNext app
CameraNext checks to see if exposure-time-values is defined, so define it to use slow-shutter feature.

Also limit exposure time to a max. of 2 seconds and a min. of 1/5000 second without returning an error to prevent the app from force closing when values greater than 2 seconds or less than 1/5000 second are selected. Note that the proprietary libraries cap exposure time to 2 seconds.

Change-Id: I79b28eb3a269d0ee2b8b453189ed4f8d69873bb7