QCamera2: Set preview size to 320x240 for torch mode
When the camera is opened, the default preview size is set to the first size in the preview-size table, aka the largest preview size available. The flash LEDs are part of the back camera, and the largest preview size for the back camera is not supported by the front camera, so when trusted-face (face unlock using the front camera) is used while the torch is enabled, the camera completely breaks and the torch gets stuck turned on until the next full reboot (as Google's trusted-face code does not expect such a large preview size).

Set the preview size for the standalone torch mode to the lowest size available (320x240) to prevent trusted-face from breaking the camera. Trusted-face and the standalone torch don't show the user the camera preview, so it's best to always set the preview size to the lowest-available resolution in this case (trusted-face also expects the lowest resolution).

Change-Id: I05892f2e74f7ee4ca45efe0adb13f52e33ba3be7