QCamera2: Fix division by 0 in mm_jpeg_exif
When the exposure time was set to a value greater than 2 seconds, the ROUND macro would round the value down to 0, causing a division by 0 and an incorrect exposure time reported in the Exif data.

Fix the division by 0 to fix the reported exposure time for exposures greater than 2 seconds.

Change-Id: I512abbc494b75e9731a0916c62ebaa53a3d154b8