QCamera2: Set preview size to 1080p for torch usecase
The Bank of America banking app's check-deposit feature opens the camera
with the torch enabled. It doesn't expect a preview size of 320x240, so
the camera subsequently crashes after opening in the BofA app.

Set the preview size to 1920x1080 for the torch usecase to prevent the
camera from crashing inside the BofA app (as well as other possible apps).
1920x1080 is the maximum resolution we can use for the torch without
invoking the trusted-face crash, so this works out perfectly.

Change-Id: I4025b003126b220737bf6a3b4f2e760af6ae5fac