QCamera2: Fix low frame rate in low lighting
Our camera backend does not cap exposure time to 1/30th of a second for videos (and
in general), resulting in very high exposure times when using the camcorder (e.g. 1/10th of a
second), as well as when taking pictures in poor lighting (resulting in motion blur).

Force the exposure time to be 1/30th of a second when the ISO is greater than 200 to
maintain a consistent 30 frames per second at all times.

Change-Id: Ieb39c9684653190b799be64ed772e5cb0440014f