QCamera2: Add 5 second delay during camera detection on boot
The camera backend probes the front-facing camera first, causing it to
sometimes inherit camera ID 0 and wreak havoc. This not-so-rare scenario
causes the rear camera to no longer be detected by Android, until the next
reboot where the bug does not manifest itself.

Therefore, add a very long delay during camera detection on boot in order to
ensure that all camera sensors are finished probing. A very long delay was
empirically determined to be necessary.

Change-Id: Ic0585a3f68bf152dd8e65e5ae2f4fcdca8a3e3a2