QCamera2: Run get_num_of_cameras() two additional times
get_num_of_cameras() is only executed on init, so the changes it makes to
g_cam_ctrl happen only at init. This means that if it populates g_cam_ctrl
with bad values on init (due to a bad run), then those bad values will
persist until the next reboot.

This is meant to address the rear camera being completely inaccessible yet
visible to Android in rare cases.

Change-Id: Ieaaaa03bb4c8b0f7ebce6ed62557f93d5019d67e